Refactored large portions of scraping code
This brings the scraping code up to speed with the revamped gevent
usage in coroutines.py and uses a ScrapeResults object which can be
added to a gevent queue. Previously, the scraping objects themselves
were added to the queue, however, they referenced attributes on self
which could have potentially caused problems when another iteration
called the object again triggering another scrape which would
overwrite those variables and lose the data. The ScrapeResults objects
store a reference to the main scraping object which allows it to
process the resulting html later via the object that inherently knows
what code to run for that particular html response.